+++ app.js
@@ -1,5 +1,5 @@
 require("dotenv").config();
-require("./src/helpers/Passport");
+require("./src/helpers/functions/Passport");
 
 const { bold } = require("colors");
 const express = require("express");
@@ -7,8 +7,8 @@ const app = express();
 const bodyParser = require("body-parser");
 const cors = require("cors");
 const Config = require("./config");
-const { sendError, generateError } = require("./src/helpers/Error");
-const { connect: connectDatabase } = require("./src/helpers/Mongoose");
+const { sendError, generateError } = require("./src/helpers/functions/Error");
+const { connect: connectDatabase } = require("./src/helpers/functions/Mongoose");
 
 connectDatabase();